src/v1/OpenX-Magma/requirements.txt
# Core ML/DL Frameworks
torch
tensorflow
tensorflow-datasets
Pillow
numpy

# Hugging Face and Model-specific Libraries
open_clip_torch
git+https://github.com/jwyang/transformers.git@dev/jwyang-v4.48.2
bitsandbytes